ML_Client/app/requirements.txt
certifi==2020.12.5
chardet==4.0.0
grpcio==1.37.1
grpcio-tools==1.37.1
idna==2.10
numpy==1.19.5
protobuf==3.15.8
pymilvus==1.0.1
requests==2.25.1
six==1.15.0
ujson==3.2.0
urllib3==1.26.4
sqlalchemy==1.4.3
pandas==1.1.5
psycopg2==2.8.6